Rename module to tracing_opencensus
