--- jaeger_tracing/README.rst
@@ -1,11 +1,0 @@
-Jaeger Tracing
-==============
-
-This addon allows to send tracing data to a Jaeger agent.
-
-Configuration
--------------
-
-The tracing is activated if the variable ``ODOO_JAEGER_TRACING`` has the value ``1`` or ``true``.
-
-A jaeger agent must be started on the same host.
