+++ tracing_opencensus/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from . import http
+from . import tracer
 from . import models
